requirements.txt
# --- Core ---
python-dotenv==1.0.1
numpy==2.1.3

# --- RAG stack ---
pinecone-client==5.0.1
sentence-transformers==2.7.0
huggingface_hub==0.25.2

# PyTorch CPU (nécessaire pour sentence-transformers)
torch==2.4.1

# --- UI / parsing ---
gradio==4.44.0
pypdf==4.3.1
markdown==3.7
Pillow==10.3.0

# --- Utils ---
tqdm==4.66.5
xxhash==3.4.1
safetensors==0.4.5

